requirements.txt
 numpy
 chatterbox-tts
 gradio
 soundfile